+++ src/main/java/domain/board/PiecesOnChessBoard.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.board;
 
+import domain.Position;
+import domain.Team;
 import domain.piece.Piece;
 import domain.piece.PieceType;
 import domain.piece.PieceWrapper;
