+++ src/main/java/domain/piece/AbstractCatchOnMovePiece.java
@@ -7,7 +7,7 @@ import static domain.PieceMoveResult.SUCCESS;
 import static domain.piece.PieceType.KING;
 
 import domain.PieceMoveResult;
-import domain.PiecesOnChessBoard;
+import domain.board.PiecesOnChessBoard;
 import domain.Position;
 import domain.Team;
 import java.util.Optional;
