+++ src/test/java/domain/piece/KnightTest.java
@@ -17,7 +17,7 @@ import static domain.Position.F5;
 import static domain.Team.BLACK;
 import static domain.Team.WHITE;
 
-import domain.PiecesOnChessBoard;
+import domain.board.PiecesOnChessBoard;
 import domain.Position;
 import java.util.Arrays;
 import java.util.List;
